requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
python-decouple>=3.8
whitenoise>=6.6.0
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
Pillow>=10.2.0
python-dotenv>=1.0.1
setuptools